ALSA: usb: fcp: Fix hwdep read ops argument
The hwdep read ops argument is long instead of ssize_t due to
historical reason.  Fix the build error.

Fixes: 46757a3e7d50 ("ALSA: FCP: Add Focusrite Control Protocol driver")
Reported-by: kernel test robot <lkp@intel.com>
Closes: https://lore.kernel.org/oe-kbuild-all/202501190536.A2CiOP0a-lkp@intel.com/
Link: https://patch.msgid.link/20250119085357.17735-1-tiwai@suse.de

+++ sound/usb/fcp.c
@@ -816,7 +816,7 @@ static int fcp_hwdep_ioctl(struct snd_hwdep *hw, struct file *file,
 }
 
 static ssize_t fcp_hwdep_read(struct snd_hwdep *hw, char __user *buf,
-			      ssize_t count, loff_t *offset)
+			      long count, loff_t *offset)
 {
 	struct usb_mixer_interface *mixer = hw->private_data;
 	struct fcp_data *private = mixer->private_data;
